[Add] add userdefaults extension
