+++ ToGather/ToGather/Screens/ToGatherApp.swift
@@ -12,6 +12,8 @@ import FirebaseCore
 struct ToGatherApp: App {
     init() {
         FirebaseApp.configure()
+        
+        // TODO: - 파이어베이스 연동과정에서 제거예정
         do {
         try UserDefaults.standard.setObject(User(id: "125",
                                        nickname: "miller",
@@ -29,41 +31,3 @@ struct ToGatherApp: App {
     }
 }
 
-// MARK: - Userdefault 커스텀 타입을 위한 코드
-
-protocol ObjectSavable {
-    func setObject<Object>(_ object: Object, forKey: String) throws where Object: Encodable
-    func getObject<Object>(forKey: String, castTo type: Object.Type) throws -> Object where Object: Decodable
-}
-extension UserDefaults: ObjectSavable {
-    func setObject<Object>(_ object: Object, forKey: String) throws where Object: Encodable {
-        let encoder = JSONEncoder()
-        do {
-            let data = try encoder.encode(object)
-            set(data, forKey: forKey)
-        } catch {
-            throw ObjectSavableError.unableToEncode
-        }
-    }
-    
-func getObject<Object>(forKey: String, castTo type: Object.Type) throws -> Object where Object: Decodable {
-        guard let data = data(forKey: forKey) else { throw ObjectSavableError.noValue }
-        let decoder = JSONDecoder()
-        do {
-            let object = try decoder.decode(type, from: data)
-            return object
-        } catch {
-            throw ObjectSavableError.unableToDecode
-        }
-    }
-}
-
-enum ObjectSavableError: String, LocalizedError {
-    case unableToEncode = "Unable to encode object into data"
-    case noValue = "No data object found for the given key"
-    case unableToDecode = "Unable to decode object into given type"
-    
-    var errorDescription: String? {
-        rawValue
-    }
-}
